# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "description": "Always accepted, whether or not the address has an account, so the response cannot be used to discover registered users. Delivery failures are logged server-side for the same reason.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Send a password reset link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "description": "Single use and time limited. On success the caller is signed in and every session issued before the change stops working.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Redeem a password reset link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "description": "Requires the current password. Other devices are signed out; this one is issued a fresh session.",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Change the signed-in account password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/billing/checkout": {
      "post": {
        "operationId": "BillingController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCheckoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Start a Polar checkout for a plan and billing interval",
        "tags": [
          "billing"
        ]
      }
    },
    "/api/billing/checkout/{checkoutId}": {
      "get": {
        "operationId": "BillingController_checkoutStatus",
        "parameters": [
          {
            "name": "checkoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Whether a completed checkout has been credited yet",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/api/candidates/{id}/renders": {
      "post": {
        "operationId": "RendersController_fromCandidate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRenderDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Generate Clip: creates a render from an immutable snapshot of the candidate range + settings.",
        "tags": [
          "renders"
        ]
      }
    },
    "/api/videos/{id}/renders": {
      "post": {
        "operationId": "RendersController_manual",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualRenderDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Manual range render (no-candidate state) or continuation of an existing lineage.",
        "tags": [
          "renders"
        ]
      },
      "get": {
        "operationId": "RendersController_forVideo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      }
    },
    "/api/renders": {
      "get": {
        "operationId": "RendersController_list",
        "parameters": [
          {
            "name": "videoId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      }
    },
    "/api/renders/{id}": {
      "get": {
        "operationId": "RendersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Render status + output; signed playback/download URLs are re-issued on every read.",
        "tags": [
          "renders"
        ]
      },
      "patch": {
        "operationId": "RendersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRenderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      },
      "delete": {
        "operationId": "RendersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      }
    },
    "/api/renders/{id}/rerender": {
      "post": {
        "operationId": "RendersController_rerender",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRenderDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      }
    },
    "/api/renders/{id}/retry": {
      "post": {
        "operationId": "RendersController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "renders"
        ]
      }
    }
  },
  "info": {
    "title": "ClipRover API",
    "description": "Product API for the ClipRover AI short-form clipping platform. All endpoints require a session cookie (cr_session) or Bearer token unless marked public. Errors use the envelope { error: { code, message, retryable, correlationId } }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cr_session"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 128
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "StartCheckoutDto": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "creator",
              "studio"
            ]
          },
          "interval": {
            "type": "string",
            "description": "Defaults to monthly. Annual is a separate Polar product per plan.",
            "enum": [
              "month",
              "year"
            ]
          }
        },
        "required": [
          "plan"
        ]
      },
...
      "CaptionsDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "preset": {
            "type": "string",
            "enum": [
              "bold-default",
              "karaoke",
              "minimal",
              "impact"
            ]
          }
        },
        "required": [
          "enabled"
        ]
      },
      "CreateRenderDto": {
        "type": "object",
        "properties": {
          "startMs": {
            "type": "number",
            "minimum": 0
          },
          "endMs": {
            "type": "number",
            "minimum": 1
          },
          "aspectRatio": {
            "enum": [
              "9:16",
              "4:5",
              "1:1",
              "16:9"
            ],
            "type": "string"
          },
          "framingMode": {
            "enum": [
              "auto",
              "center",
              "fit"
            ],
            "type": "string"
          },
          "captions": {
            "$ref": "#/components/schemas/CaptionsDto"
          },
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "force": {
            "type": "boolean",
            "description": "Render again even if an identical completed render exists."
          }
        }
      },
      "ManualRenderDto": {
        "type": "object",
        "properties": {
          "startMs": {
            "type": "number",
            "minimum": 0
          },
          "endMs": {
            "type": "number",
            "minimum": 1
          },
          "aspectRatio": {
            "enum": [
              "9:16",
              "4:5",
              "1:1",
              "16:9"
            ],
            "type": "string"
          },
          "framingMode": {
            "enum": [
              "auto",
              "center",
              "fit"
            ],
            "type": "string"
          },
          "captions": {
            "$ref": "#/components/schemas/CaptionsDto"
          },
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "force": {
            "type": "boolean",
            "description": "Render again even if an identical completed render exists."
          },
          "parentRenderId": {
            "type": "string",
            "description": "Continue an existing render lineage (version history) instead of starting a new one."
          }
        }
      },
      "UpdateRenderDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "title"
        ]
      }
    }
  }
}